# Split JS files and optimise Gradle build