--- a/Instructions.md
+++ b/Instructions.md
@@ -48,7 +48,6 @@ Firstly, we need to setup the Python side:
 
 Now, let's setup the Gradle side:
 1. Import the `build.gradle` into your IDE
-    - Due to the old Gradle 5.6.4 in use here, you'll need to use Java 8 for it to work
 2. Run `gradlew setupTest`
 
 ## Step 2: Run PageGen