--- a/build.gradle
+++ b/build.gradle
@@ -72,7 +72,7 @@ void writeArtifact(group, name, version, classifier = null, String ext = 'jar') 
 }
 
 void writeXml(String group, String name, List vers) {
-    def writer = new StringWriter()
+    final writer = new StringWriter()
     def xml = new groovy.xml.MarkupBuilder(writer)
     String path = "./test/maven/${group}/${name}"
     
@@ -102,7 +102,7 @@ void writeXml(String group, String name, List vers) {
             adfocus: "271228"
     ]).toPrettyString()
     file("${path}/promotions_slim.json").text = new groovy.json.JsonBuilder([
-        homepage: 'https://not_real/',
+        homepage: 'https://not_real.invalid/',
         promos: [
             '1.16.5-latest':      '36.1.2',
             '1.16.5-recommended': '36.1.2',