--- a/requirements.txt
+++ b/requirements.txt
@@ -1,4 +1,4 @@
 Jinja2==2.11.3
-Markdown==3.3.4
+Markdown==3.3.7
 MarkupSafe==1.1.1
-requests~=2.25.1
+requests~=2.28.1