--- a/templates/page_footer.html
+++ b/templates/page_footer.html
@@ -16,9 +16,9 @@
 <script src="https://cdnjs.cloudflare.com/ajax/libs/jScrollPane/2.0.23/script/jquery.jscrollpane.min.js" crossorigin="anonymous"></script>
 <script src="https://cdnjs.cloudflare.com/ajax/libs/tooltipster/3.3.0/js/jquery.tooltipster.min.js" crossorigin="anonymous"></script>
 <script src="{{ static_root }}js/theme-switch-toggle.min.js?{{ now.timestamp() }}"></script>
-<script src="{{ static_root }}js/curse-ads.min.js?{{ now.timestamp() }}"></script>
-<script src="{{ static_root }}js/sidebar.min.js?{{ now.timestamp() }}"></script>
-<script src="{{ static_root }}js/files.min.js?{{ now.timestamp() }}"></script>
+<script src="{{ static_root }}js/curse-ads.min.js?1"></script> <!-- manually increment the ?1 when you change the file -->
+<script src="{{ static_root }}js/sidebar.min.js?1"></script>
+<script src="{{ static_root }}js/files.min.js?1"></script>
 {% if artifact and artifact.config['analytics'] %}{{ artifact.config['analytics'] }}{% endif %}
 
 {%- if artifact and artifact.config['body_end'] %}