--- a/templates/page_footer.html
+++ b/templates/page_footer.html
@@ -11,10 +11,10 @@
         </div>
     </footer>
 </div>
-<script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/2.2.4/jquery.min.js" crossorigin="anonymous"></script>
+<script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.6.3/jquery.min.js"></script>
 <script src="https://use.fontawesome.com/a1f20be65b.js"></script>
-<script src="https://cdnjs.cloudflare.com/ajax/libs/jScrollPane/2.0.23/script/jquery.jscrollpane.min.js" crossorigin="anonymous"></script>
-<script src="https://cdnjs.cloudflare.com/ajax/libs/tooltipster/3.3.0/js/jquery.tooltipster.min.js" crossorigin="anonymous"></script>
+<script src="https://cdnjs.cloudflare.com/ajax/libs/jScrollPane/2.0.23/script/jquery.jscrollpane.min.js"></script>
+<script src="https://cdnjs.cloudflare.com/ajax/libs/tooltipster/3.3.0/js/jquery.tooltipster.min.js"></script>
 <script src="{{ static_root }}js/theme-switch-toggle.min.js?{{ crc32(static_root + 'js/theme-switch-toggle.min.js') }}"></script>
 <script src="{{ static_root }}js/curse-ads.min.js?{{ crc32(static_root + 'js/curse-ads.min.js') }}"></script>
 <script src="{{ static_root }}js/sidebar.min.js?{{ crc32(static_root + 'js/sidebar.min.js') }}"></script>