--- a/templates/page_header.html
+++ b/templates/page_header.html
@@ -41,7 +41,7 @@
     <link rel="stylesheet" data-theme="dark" href="{{ static_root }}css/styles_dark.css?{{ crc32(static_root + 'css/styles_dark.css') }}">
     <script src="{{ static_root }}js/theme-switch.min.js?{{ crc32(static_root + 'js/theme-switch.min.js') }}"></script>
     <!-- relative time in browser -->
-    <script type="text/javascript" src="https://cdn.jsdelivr.net/combine/npm/dayjs@1.10.4,npm/dayjs@1.10.4/plugin/relativeTime.min.js"></script>
+    <script src="https://cdn.jsdelivr.net/combine/npm/dayjs@1.10.8,npm/dayjs@1.10.8/plugin/relativeTime.min.js"></script>
     <script>dayjs.extend(dayjs_plugin_relativeTime)</script>
 
     <!-- HTML5 shim and Respond.js for IE8 support of HTML5 elements and media queries -->