# Summary
- Split up JS to help simplify the move to a future full-Groovy system and allow for async JS optimisations later on 
- Used [Gradle's Task Configuration Avoidance API](https://docs.gradle.org/current/userguide/task_configuration_avoidance.html) to speed up builds
- Changed jQuery link to a non-Google one that should work in China
- Use CRC32 checksums for cache busting instead of the time of the generated page

# Details and rationale

## Split JS files
JS files used to be merged because at the time, SPDY and HTTP/2 weren't commonplace and merging files to reduce sequential requests tends to reduce overhead and improve performance on HTTP/1.

This is no longer necessary in HTTP/2 (file downloads are multiplexed, so head-of-line blocking is no longer an issue that needs manual workarounds like merging) and causes all our JS files to be redownloaded in full if a change is made to any single file, which hurts cache hit rates.

Splitting up the JS will simplify the move to a future full-Groovy build system and make more optimisations possible later on (such as having different JS files download and execute asynchronously).

### jQuery URL change
After seeing #24, I thought it would make sense to change the jQuery URL to a non-Google one so that it can be accessed in China, resulting in a less broken site for them.

### CRC32 checksum cache busting
Currently the files site uses the timestamp of when the page was generated for the JS files, even if they haven't changed. This forces the browser to redownload these files every time a new version is released.

This has been changed for a CRC32 checksum system that does not change if the file remains the same.

## Optimise Gradle build
All task definitions have been changed to use [Gradle's Task Configuration Avoidance API](https://docs.gradle.org/current/userguide/task_configuration_avoidance.html) to speed up builds. The Gradle part of the build system is now noticeably faster for the "configuring tasks" stage as it no longer needs to eagerly create any tasks at configuration time.

I've also explicitly typed a few more things here and there to speed up the "compile Groovy buildscript" stage slightly.

# Benchmarks

## Gradle
Deleted the `.gradle` and `build` folders, then ran `gradlew --scan bundleFiles`

Before: https://scans.gradle.com/s/is6p2xlruxak4/performance/configuration
After: https://scans.gradle.com/s/wkuvdn6w3rzxw/performance/configuration

| Metric | Before | After |
|--------|---------|-------|
| # of tasks created immediately | 17 | 8 |
| Model configuration | 2.9s | 0.4s |
| Task graph configuration | 0.1s | 0.003s |
| Total configuration time | 4.6s | 0.4s |

## Site
`gradlew setupTest` then `gradlew runTestPageGen`, copied over each test build over to its own folder and fixed file links as necessary. F12 network tab simulating fast 3G with cache disabled. Tested twice - once with XAMPP (HTTP/1) and once with [simplehttp2server](https://github.com/GoogleChromeLabs/simplehttp2server) (HTTP/2)

For HTTP/1, page load was about 0.5-1s slower on avg.
For HTTP/2, it was within margin of error for an uncached page load.

The performance benefit becomes a bit more apparent when a script is changed. For example, if  `files.js`  is changed it'll now only redownload that one file rather than the entire merged js. More performance gains are now possible with some changes to the JS, allowing different scripts to be loaded async in the future.

Considering that the aim of this PR wasn't to improve page load times on its own (but instead to help simplify the move to newer things and allow for optimisations in future PRs), it's good to know that performance at least isn't any worse over HTTP/2.