@LexManos 2022-07-11
`<!-- manually increment the ?1 when you change the file -->` this is a no go. Is it possible to make this cache buster the CRC of the js files or something else automatic?