@PaintNinja 2022-10-16
Had some trouble figuring out how to get the file path from the template to CRC, so I've set it to be the year and week number instead of the precise datetime down to the millisecond of the build.

While this isn't ideal, it's much better having the cache reused for a minimum of a week rather than redownloading for every new Forge build.