@LexManos 2022-10-16
No it's not acceptable to wait a week to fix one line typos or test changes. That's why I had the timestamp in the first place because I got annoied while testing changes and not having them show up.