@PaintNinja 2022-12-17
Okay, should be using CRC checksums now