@LexManos 2023-02-01
After readding the google features which were removed for no good reason, i'll pull this.
We'll see if it breaks anything.